refactor(memory,controlcenter): restore memory purity + drop handler_tiers hot-path scan (#341)
- Move ConsolidatePreferences out of internal/memory/ into
  internal/memory/curation/preferences.go so memory/ no longer imports
  ai/provider (ARCHITECTURE-v0.7.10.md §2.2). Consolidator now takes a
  curation.PrefInvoker closure, wired at the single prod call-site
  (comms/reaction.go) without creating a new cross-package adapter.
- Export memory.PreferencesFile / memory.PreferencesThreshold /
  memory.LockPreferences/UnlockPreferences so curation can share the
  file mutex and constants without duplication.
- Drop the tooling.NewRegistry(h.DataDir) fallback in
  controlcenter/handler_tiers.go — every GET /api/tiers was re-scanning
  tools.d/*.json on disk when ToolRegistry was unexpectedly nil.
  The daemon bootstrap always wires it; fallback is dead code in prod.
- Move 5 preference-consolidation tests to curation/preferences_test.go
  (kept the black-box package split, switched stubProvider to a
  PrefInvoker-shaped stubInvoker).

make regression green. Global coverage 48.2% (floor per #343: 47.0%).
Archtest TestFoundationDependencyRules / TestConsumerDependencyRules /
TestHardcodedModelFallback all enforcing-mode PASS.
